# Fix performance problems with new Madelung method