--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -80,7 +80,6 @@ add_subdirectory(TotalEnergy)
 
 add_subdirectory(VORPOL)
 
-
 target_link_libraries(lsms_main PUBLIC lsmscore)
 
 target_link_libraries(lsms_wl PUBLIC lsmscore)