--- a/src/MultipoleMadelung/CMakeLists.txt
+++ b/src/MultipoleMadelung/CMakeLists.txt
@@ -16,4 +16,7 @@ target_sources(lsmscore PUBLIC
         debug.hpp
         common.hpp)
 
+
+
+
 target_include_directories(lsmscore PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})