Problem
=======

There are significant  timing difference between the old and new Madelung code for large scale calculations:

   8192 atoms: old:  0.40s new:  12.85s
  65536 atoms: old:  3.16s new: 120.53s
 524288 atoms: old: 24.63s new: 953.56s

Fix
===

This problem was caused by too less optimization steps for finding the optimal scaling of the lattice.
I fixed the problem and also add some improvements.

- The number of optimization steps were increased
- Bisection for finding the scaling parameter
- Adding OpenMP parallelism.